代码更新

+++ mmall/src/main/java/com/mmall/model/po/Cart.java
@@ -14,8 +14,8 @@ public class Cart {
 
     private Integer checked;
 
-    private Long createTime;
+    private Integer createTime;
 
-    private Long updateTime;
+    private Integer updateTime;
 
 }
