+++ mmall/src/main/java/com/mmall/model/po/OrderItem.java
@@ -24,8 +24,8 @@ public class OrderItem {
 
     private BigDecimal totalPrice;
 
-    private Long createTime;
+    private Integer createTime;
 
-    private Long updateTime;
+    private Integer updateTime;
 
 }
